Move every gui listener into event/listener
